fix:sponser section styled,files renamed
