+++ config/content/Sponsors.js
@@ -55,6 +55,41 @@ export default {
       alt: 'Auth0',
     },
     {
+      src: 'https://res.cloudinary.com/riteshsp2000/image/upload/project-oregano/sponsors/beta/lr46yhbzdbtmv4vdm9yg.png',
+      name: 'Egghead',
+      size: '100%',
+      link: 'https://egghead.io/',
+      width: '200px',
+      alt: 'Egghead',
+      margin: '0px',
+    },
+    {
+      src: 'https://res.cloudinary.com/riteshsp2000/image/upload/v1634795359/project-oregano/sponsors/Group_1968_gb9rmx.png',
+      name: 'Codedamn',
+      size: '100%',
+      link: 'https://codedamn.com/',
+      width: '200px',
+      alt: 'Codedamn',
+      margin: '0px',
+    },
+    {
+      src: 'https://res.cloudinary.com/riteshsp2000/image/upload/project-oregano/sponsors/beta/iylv4cyswfjzoydb3why.png',
+      name: 'Sashido',
+      size: '98%',
+      link: 'https://www.sashido.io/en/',
+      width: '200px',
+      alt: 'Sashido',
+    },
+    {
+      src: 'https://res.cloudinary.com/riteshsp2000/image/upload/project-oregano/sponsors/beta/zuvuxsxn7zukvrtx5swz.png',
+      name: '.xyz',
+      size: '70%',
+      link: 'https://gen.xyz/',
+      width: '130px',
+      alt: '.xyz',
+      margin: '0px',
+    },
+    {
       src: 'https://res.cloudinary.com/riteshsp2000/image/upload/project-oregano/sponsors/beta/ogjutyp26djnh9lrzybl.png',
       name: 'Celo',
       size: '95%',
@@ -87,47 +122,12 @@ export default {
       alt: 'Tezos',
     },
     {
-      src: 'https://res.cloudinary.com/riteshsp2000/image/upload/project-oregano/sponsors/beta/iylv4cyswfjzoydb3why.png',
-      name: 'Sashido',
-      size: '98%',
-      link: 'https://www.sashido.io/en/',
-      width: '200px',
-      alt: 'Sashido',
-    },
-    {
-      src: 'https://res.cloudinary.com/riteshsp2000/image/upload/project-oregano/sponsors/beta/zuvuxsxn7zukvrtx5swz.png',
-      name: '.xyz',
-      size: '70%',
-      link: 'https://gen.xyz/',
-      width: '130px',
-      alt: '.xyz',
-      margin: '0px',
-    },
-    {
       src: 'https://res.cloudinary.com/riteshsp2000/image/upload/project-oregano/sponsors/beta/ps74duvfnma7sjh455de.svg',
       name: 'Filecoin',
       size: '100%',
       link: 'https://filecoin.io/',
       width: '200px',
       alt: 'Filecoin',
-      margin: '0px',
-    },
-    {
-      src: 'https://res.cloudinary.com/riteshsp2000/image/upload/project-oregano/sponsors/beta/lr46yhbzdbtmv4vdm9yg.png',
-      name: 'Egghead',
-      size: '100%',
-      link: 'https://egghead.io/',
-      width: '200px',
-      alt: 'Egghead',
-      margin: '0px',
-    },
-    {
-      src: 'https://res.cloudinary.com/riteshsp2000/image/upload/v1634795359/project-oregano/sponsors/Group_1968_gb9rmx.png',
-      name: 'Codedamn',
-      size: '100%',
-      link: 'https://codedamn.com/',
-      width: '200px',
-      alt: 'Codedamn',
       margin: '0px',
     },
   ],
@@ -192,7 +192,7 @@ export default {
       alt: 'Tweetbeam',
     },
     {
-      src: 'https://res.cloudinary.com/riteshsp2000/image/upload/project-oregano/sponsors/gamma/hendtn2wfa16wyuxk7.png',
+      src: 'https://res.cloudinary.com/riteshsp2000/image/upload/v1634369574/project-oregano/sponsors/gamma/hendtn2wfa16wyuxk7.png',
       name: 'Voiceflow',
       size: '100%',
       link: 'https://www.voiceflow.com/',
