+++ src/components/CommunitySection/Card.jsx
