+++ src/components/CommunitySection/Carousel.jsx
@@ -43,7 +43,7 @@ const Carousel = ({ colors, cards }) => {
     arrows: false,
     slidesToScroll: 1,
     autoplay: true,
-    autoplaySpeed: 2000,
+    autoplaySpeed: 1500,
     cssEase: 'ease',
     responsive: [
       {
