+++ src/components/CommunitySection/carousel.css
