+++ src/components/SponsorsSection/Sponsor.jsx
@@ -56,14 +56,14 @@ function Sponsor() {
       title: 'MENTAL HEALTH PARTNER',
       sponsors: sponsors['mental-health-partner'],
       bcolor: 'transparent',
-      span: 'span 3',
+      span: 'span 2',
       grid: '1fr 1fr 1fr ',
     },
     {
       title: 'CODING PARTNER',
       sponsors: sponsors['coding-partner'],
       bcolor: 'transparent',
-      span: 'span 2',
+      span: 'span 1',
       grid: '1fr 1fr 1fr ',
     },
   ];
